+++ CHANGELOG.md
@@ -26,7 +26,6 @@
 - `Innmind\Xml\Node\ProcessingInformation` is now internal, use `Innmind\Xml\Node::processingInformation()` instead
 - `Innmind\Xml\Node\Text` is now internal, use `Innmind\Xml\Node::text()` instead
 - `Innmind\Xml\Reader` is now a final class
-- `Innmind\Xml\Attribute` always render its value even if it's empty
 
 ### Removed
 
@@ -39,6 +38,7 @@
 - `Innmind\Xml\Node::appendChild()`
 - `Innmind\Xml\AsContent`
 - `Innmind\Xml\Attribute::empty()`
+- `Innmind\Xml\Attribute::toString()`
 
 ### Fixed